requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file
Flask>=2.0.1
gunicorn>=20.0.0
flask-cors>=3.0.10
openai>=1.0.0
python-dotenv>=0.15.0
Mangum>=0.10.0